changed prototype for _getlong

+++ src/res.c
@@ -4,7 +4,7 @@
  * shape or form. The author takes no responsibility for any damage or loss
  * of property which results from the use of this software.
  *
- * $Id: res.c,v 1.37 1999/07/19 01:00:33 tomh Exp $
+ * $Id: res.c,v 1.38 1999/07/19 01:09:49 tomh Exp $
  *
  * July 1999 - Rewrote a bunch of stuff here. Change hostent builder code,
  *     added callbacks and reference counting of returned hostents.
@@ -255,7 +255,7 @@ static  struct  resinfo {
  *
  */
 extern u_short  _getshort(const u_char *);
-extern u_long   _getlong(const u_char *);
+extern u_int    _getlong(const u_char *);
 /*
  * int
  * res_isourserver(ina)
